refactor: move components into relevant folders for organisation
