+++ src/components/dashboard/BedTiles.tsx
