+++ src/pages/wards.tsx
@@ -10,9 +10,9 @@ import {
 import { fetchBedByBedId } from "./api/smartbed_api";
 import autoAnimate from "@formkit/auto-animate";
 import profilePic from "../../public/profilepic.png";
-import VitalTiles from "@/components/VitalTiles";
-import TileCustomisationModal from "@/components/TileCustomisationModal";
-import BedTiles from "@/components/BedTiles";
+import VitalTiles from "@/components/dashboard/VitalTiles";
+import TileCustomisationModal from "@/components/dashboard/TileCustomisationModal";
+import BedTiles from "@/components/dashboard/BedTiles";
 import ReportProblemIcon from "@mui/icons-material/ReportProblem";
 import { VirtualNurse } from "@/types/virtualNurse";
 import { Link, Tooltip } from "@mui/material";
@@ -20,7 +20,7 @@ import { Ward } from "@/types/ward";
 import { SocketContext } from "@/pages/layout";
 import { Alert } from "@/types/alert";
 import { Patient } from "@/types/patient";
-import DashboardAlertIcon from "@/components/DashboardAlertIcon";
+import DashboardAlertIcon from "@/components/dashboard/DashboardAlertIcon";
 import { fetchAlertsByPatientId } from "./api/patients_api";
 import GridViewIcon from "@mui/icons-material/GridView";
 import ListIcon from "@mui/icons-material/List";
